+++ src/components/layouts/DashBoard.tsx
@@ -114,7 +114,7 @@ export default function DashBoard({ children }: LayoutProps) {
             boxSizing: "border-box",
           },
         }}
-        variant="persistent"
+        variant="temporary"
         anchor="left"
         open={open}
       >
