+++ src/features/audioPlayer/components/SoundControlBar.tsx
@@ -1,116 +1,21 @@
-import * as React from "react";
-import { styled, useTheme } from "@mui/material/styles";
-import Box from "@mui/material/Box";
-import Typography from "@mui/material/Typography";
-import Slider from "@mui/material/Slider";
 import VolumeSlider from "./VolumeSlider";
+import Box from "@mui/material/Box";
 import PlaySkipControls from "./PlaySkipControls";
+import PlayBackBar from "./PlayBackBar";
+import TrackCard from "./TrackCard";
 
-// const Widget = styled("div")(({ theme }) => ({
-//   padding: 16,
-//   borderRadius: 16,
-//   // width: 343,
-//   // maxWidth: "100%",
-//   margin: "auto",
-//   position: "relative",
-//   zIndex: 1,
-//   backgroundColor:
-//     theme.palette.mode === "dark" ? "rgba(0,0,0,0.6)" : "rgba(255,255,255,0.4)",
-//   backdropFilter: "blur(40px)",
-// }));
-
-const CoverImage = styled("div")({
-  width: 100,
-  height: 100,
-  objectFit: "cover",
-  overflow: "hidden",
-  flexShrink: 0,
-  borderRadius: 8,
-  backgroundColor: "rgba(0,0,0,0.08)",
-  "& > img": {
-    width: "100%",
-  },
-});
-
-const TinyText = styled(Typography)({
-  fontSize: "0.75rem",
-  opacity: 0.38,
-  fontWeight: 500,
-  letterSpacing: 0.2,
-});
-
-export default function MusicPlayerSlider() {
-  const theme = useTheme();
-  const duration = 200; // seconds
-  const [position, setPosition] = React.useState(32);
-
-  function formatDuration(value: number) {
-    const minute = Math.floor(value / 60);
-    const secondLeft = value - minute * 60;
-    return `${minute}:${secondLeft < 10 ? `0${secondLeft}` : secondLeft}`;
-  }
-
+export default function SoundControlBar() {
   return (
-    <Box sx={{ width: "100%", overflow: "hidden" }}>
-      <Box sx={{ display: "flex", alignItems: "center" }}>
-        <Box sx={{ ml: 1.5, minWidth: 0 }}>
-          <Typography variant="caption" color="text.secondary" fontWeight={500}>
-            Jun Pulse
-          </Typography>
-          <Typography noWrap>
-            <b>คนเก่าเขาทำไว้ดี (Can&apos;t win)</b>
-          </Typography>
-          <Typography noWrap letterSpacing={-0.25}>
-            Chilling Sunday &mdash; คนเก่าเขาทำไว้ดี
-          </Typography>
-        </Box>
-      </Box>
-      <Slider
-        aria-label="time-indicator"
-        size="small"
-        value={position}
-        min={0}
-        step={1}
-        max={duration}
-        onChange={(_, value) => setPosition(value as number)}
-        sx={{
-          color: theme.palette.mode === "dark" ? "#fff" : "rgba(0,0,0,0.87)",
-          height: 4,
-          "& .MuiSlider-thumb": {
-            width: 8,
-            height: 8,
-            transition: "0.3s cubic-bezier(.47,1.64,.41,.8)",
-            "&:before": {
-              boxShadow: "0 2px 12px 0 rgba(0,0,0,0.4)",
-            },
-            "&:hover, &.Mui-focusVisible": {
-              boxShadow: `0px 0px 0px 8px ${
-                theme.palette.mode === "dark"
-                  ? "rgb(255 255 255 / 16%)"
-                  : "rgb(0 0 0 / 16%)"
-              }`,
-            },
-            "&.Mui-active": {
-              width: 20,
-              height: 20,
-            },
-          },
-          "& .MuiSlider-rail": {
-            opacity: 0.28,
-          },
-        }}
-      />
-      <Box
-        sx={{
-          display: "flex",
-          alignItems: "center",
-          justifyContent: "space-between",
-          mt: -2,
-        }}
-      >
-        <TinyText>{formatDuration(position)}</TinyText>
-        <TinyText>-{formatDuration(duration - position)}</TinyText>
-      </Box>
+    <Box
+      sx={{
+        width: "100vw",
+        backgroundColor: "red",
+        position: "fixed",
+        bottom: 0,
+      }}
+    >
+      <TrackCard />
+      <PlayBackBar />
       <PlaySkipControls />
       <VolumeSlider />
     </Box>
