+++ src/pages/index.tsx
@@ -1,6 +1,7 @@
 import { Inter } from "next/font/google";
 import DashBoard from "@/components/layouts/DashBoard";
 import Audio from "@/features/audioList/components/Audio";
+import SoundControlBar from "@/features/audioPlayer/components/SoundControlBar";
 
 const inter = Inter({ subsets: ["latin"] });
 
@@ -9,6 +10,7 @@ export default function Home() {
     <main>
       <DashBoard>
         <Audio />
+        <SoundControlBar />
       </DashBoard>
     </main>
   );
